+++ backend/canisters/group_index/api/src/lifecycle/init.rs
@@ -8,6 +8,7 @@ pub struct Args {
     pub service_principals: Vec<Principal>,
 
     pub group_canister_wasm: CanisterWasm,
+    pub community_canister_wasm: CanisterWasm,
     pub local_group_index_canister_wasm: CanisterWasm,
     pub user_index_canister_id: CanisterId,
     pub cycles_dispenser_canister_id: CanisterId,