+++ backend/canisters/local_group_index/api/src/updates/c2c_create_community.rs
@@ -1,0 +1,29 @@
+use candid::{CandidType, Principal};
+use serde::{Deserialize, Serialize};
+use types::{Avatar, CommunityId, CommunityPermissions, GroupGate, GroupRules, UserId};
+
+#[derive(CandidType, Serialize, Deserialize, Debug)]
+pub struct Args {
+    pub created_by_user_id: UserId,
+    pub created_by_user_principal: Principal,
+    pub is_public: bool,
+    pub name: String,
+    pub description: String,
+    pub rules: GroupRules,
+    pub avatar: Option<Avatar>,
+    pub history_visible_to_new_joiners: bool,
+    pub permissions: Option<CommunityPermissions>,
+    pub gate: Option<GroupGate>,
+}
+
+#[derive(CandidType, Serialize, Deserialize, Debug)]
+pub enum Response {
+    Success(SuccessResult),
+    CyclesBalanceTooLow,
+    InternalError(String),
+}
+
+#[derive(CandidType, Serialize, Deserialize, Debug)]
+pub struct SuccessResult {
+    pub community_id: CommunityId,
+}