+++ backend/canisters/local_group_index/api/src/updates/c2c_delete_community.rs
@@ -1,13 +1,14 @@
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
-use types::ChatId;
+use types::CommunityId;
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
-    pub group_id: ChatId,
+    pub community_id: CommunityId,
 }
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub enum Response {
     Success,
+    CommunityNotFound,
 }