+++ backend/canisters/local_group_index/api/src/updates/c2c_upgrade_community_canister_wasm.rs
@@ -1,0 +1,11 @@
+use candid::CandidType;
+use serde::{Deserialize, Serialize};
+use types::UpgradeCanisterWasmArgs;
+
+pub type Args = UpgradeCanisterWasmArgs;
+
+#[derive(CandidType, Serialize, Deserialize, Debug)]
+pub enum Response {
+    Success,
+    VersionNotHigher,
+}