+++ backend/integration_tests/src/wasms.rs
@@ -5,6 +5,7 @@ use std::io::Read;
 use types::{CanisterWasm, Version};
 
 lazy_static! {
+    pub static ref COMMUNITY: CanisterWasm = get_canister_wasm("community");
     pub static ref CYCLES_DISPENSER: CanisterWasm = get_canister_wasm("cycles_dispenser");
     pub static ref CYCLES_MINTING_CANISTER: CanisterWasm = get_canister_wasm("cycles_minting_canister");
     pub static ref GROUP: CanisterWasm = get_canister_wasm("group");