+++ backend/canisters/local_group_index/CHANGELOG.md
@@ -5,9 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Added
+
+- Re-introduce invite by code on backend ([#3552](https://github.com/open-ic/open-chat/pull/3552))
+
 ## [[2.0.664](https://github.com/open-ic/open-chat/releases/tag/v2.0.664-local_group_index)] - 2023-04-17
 
-# Added
+### Added
 
 - Implement 'Gated Groups' ([#3406](https://github.com/open-ic/open-chat/pull/3406))